requirements.txt
setuptools~=45.2.0
opencv-python~=4.5.5.64
imutils~=0.5.4
numpy~=1.22.3
Pillow~=9.1.0
pytesseract~=0.3.9